requirements.txt
python-dotenv==1.0.0
gql==3.4.1
aiohttp==3.9.1
requests==2.31.0 